feat: 로그아웃 confirm 메세지 추가

+++ index.js
@@ -44,6 +44,10 @@ function displayUsername() {
 }
 
 function logout() {
-  window.localStorage.clear();
-  location.reload();
+  const result = window.confirm('로그아웃 하시겠습니까?');
+
+  if (result) {
+    window.localStorage.clear();
+    location.reload();
+  }
 }
